chore: reusing utility
