+++ lgn-messages/src/types/v1/preprocessing/mod.rs
@@ -1,6 +1,8 @@
 use alloy_primitives::Address;
 use alloy_primitives::U256;
+use anyhow::bail;
 use ethers::prelude::H256;
+use ethers::utils::rlp::Rlp;
 use mp2_common::digest::TableDimension;
 use mp2_v1::values_extraction;
 use serde_derive::Deserialize;
@@ -26,6 +28,47 @@ pub mod db_tasks;
 pub mod ext_tasks;
 
 const KEYS_PREPROCESSING_PREFIX: &str = "V1_PREPROCESSING";
+
+/// Different types of node types.
+#[derive(Debug, PartialEq, Eq)]
+pub enum NodeType {
+    Branch,
+    Extension,
+    Leaf,
+}
+
+/// Returns the node type given an encoded node.
+///
+/// The node spec is at [1].
+///
+/// 1- https://github.com/ethereum/execution-specs/blob/78fb726158c69d8fa164e28f195fabf6ab59b915/src/ethereum/cancun/trie.py#L177-L191
+pub fn node_type(rlp_data: &[u8]) -> anyhow::Result<NodeType> {
+    let rlp = Rlp::new(rlp_data);
+
+    let item_count = rlp.item_count()?;
+
+    if item_count == 17 {
+        Ok(NodeType::Branch)
+    } else if item_count == 2 {
+        // The first item is the encoded path, if it begins with a 2 or 3 it is a leaf, else it is
+        // an extension node
+        let first_item = rlp.at(0)?;
+
+        // We want the first byte
+        let first_byte = first_item.as_raw()[0];
+
+        // The we divide by 16 to get the first nibble
+        match first_byte / 16 {
+            0 | 1 => Ok(NodeType::Extension),
+            2 | 3 => Ok(NodeType::Leaf),
+            _ => {
+                bail!("Expected compact encoding beginning with 0,1,2 or 3")
+            },
+        }
+    } else {
+        bail!("RLP encoded Node item count was {item_count}, expected either 17 or 2")
+    }
+}
 
 #[derive(Deserialize, Serialize)]
 pub struct WorkerTask {
